Move terminal formatter
