+++ lib/wake-utils/terminal_formatter.rb
